code clean up and test improvements
