+++ schema/shared/src/test/scala-3/zio/blocks/schema/migration/CompileTimeValidationSpec.scala
@@ -3,1108 +3,21 @@ package zio.blocks.schema.migration
 import zio.test._
 import zio.test.Assertion
 import zio.blocks.schema._
-import zio.blocks.schema.migration.ShapeExtraction._
 import zio.blocks.schema.CompanionOptics
 
-object CompileTimeValidationSpec extends ZIOSpecDefault {
-
-  // Test case classes - identical schemas
-  case class PersonA(name: String, age: Int)
-  case class PersonB(name: String, age: Int)
-
-  // Test case classes - field renamed
-  case class RenameSource(oldName: String, age: Int)
-  case class RenameTarget(newName: String, age: Int)
-
-  // Test case classes - field dropped
-  case class DropSource(name: String, age: Int, extra: Boolean)
-  case class DropTarget(name: String, age: Int)
-
-  // Test case classes - field added
-  case class AddSource(name: String, age: Int)
-  case class AddTarget(name: String, age: Int, extra: Boolean)
-
-  // Test case classes - complex: drop + add + rename
-  case class ComplexSource(a: String, b: Int, c: Boolean, d: Double)
-  case class ComplexTarget(x: String, b: Int, y: Boolean, e: Long)
-
-  // Test case classes - many shared fields
-  case class ManySharedSource(shared1: String, shared2: Int, shared3: Boolean, removed: Double)
-  case class ManySharedTarget(shared1: String, shared2: Int, shared3: Boolean, added: Long)
-
-  // Test case classes - for joinFields/splitField tracking
-  case class FullNameSource(firstName: String, lastName: String, age: Int)
-  case class FullNameTarget(fullName: String, age: Int)
-
-  // Schemas
-  implicit val personASchema: Schema[PersonA]                   = Schema.derived
-  implicit val personBSchema: Schema[PersonB]                   = Schema.derived
-  implicit val renameSourceSchema: Schema[RenameSource]         = Schema.derived
-  implicit val renameTargetSchema: Schema[RenameTarget]         = Schema.derived
-  implicit val dropSourceSchema: Schema[DropSource]             = Schema.derived
-  implicit val dropTargetSchema: Schema[DropTarget]             = Schema.derived
-  implicit val addSourceSchema: Schema[AddSource]               = Schema.derived
-  implicit val addTargetSchema: Schema[AddTarget]               = Schema.derived
-  implicit val complexSourceSchema: Schema[ComplexSource]       = Schema.derived
-  implicit val complexTargetSchema: Schema[ComplexTarget]       = Schema.derived
-  implicit val manySharedSourceSchema: Schema[ManySharedSource] = Schema.derived
-  implicit val manySharedTargetSchema: Schema[ManySharedTarget] = Schema.derived
-  implicit val fullNameSourceSchema: Schema[FullNameSource]     = Schema.derived
-  implicit val fullNameTargetSchema: Schema[FullNameTarget]     = Schema.derived
-
-  def spec = suite("CompileTimeValidationSpec")(
-    suite("build method")(
-      test("build compiles for identical schemas") {
-        val migration = MigrationBuilder
-          .newBuilder[PersonA, PersonB]
-          .build
-
-        val personA = PersonA("John", 30)
-        val result  = migration(personA)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name == "John").getOrElse(false))
-      },
-      test("build compiles for complete drop migration") {
-        val migration = MigrationBuilder
-          .newBuilder[DropSource, DropTarget]
-          .dropField(_.extra, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          .build
-
-        val source = DropSource("John", 30, true)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("John")) &&
-        assertTrue(result.map(_.age) == Right(30))
-      },
-      test("build compiles for complete add migration") {
-        val migration = MigrationBuilder
-          .newBuilder[AddSource, AddTarget]
-          .addField(_.extra, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          .build
-
-        val source = AddSource("John", 30)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("John")) &&
-        assertTrue(result.map(_.age) == Right(30)) &&
-        assertTrue(result.map(_.extra) == Right(false))
-      },
-      test("build compiles for complete rename migration") {
-        val migration = MigrationBuilder
-          .newBuilder[RenameSource, RenameTarget]
-          .renameField(_.oldName, _.newName)
-          .build
-
-        val source = RenameSource("John", 30)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.newName) == Right("John")) &&
-        assertTrue(result.map(_.age) == Right(30))
-      },
-      test("build compiles for complex migration with multiple operations") {
-        // ComplexSource: a, b, c, d
-        // ComplexTarget: x, b, y, e
-        // shared: b
-        // removed: a, c, d (need to handle)
-        // added: x, y, e (need to provide)
-
-        val migration = MigrationBuilder
-          .newBuilder[ComplexSource, ComplexTarget]
-          .renameField(_.a, _.x)                                                        // handles a, provides x
-          .renameField(_.c, _.y)                                                        // handles c, provides y
-          .dropField(_.d, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double)) // handles d
-          .addField(_.e, SchemaExpr.Literal[DynamicValue, Long](0L, Schema.long))       // provides e
-          .build
-
-        val source = ComplexSource("hello", 42, true, 3.14)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.x) == Right("hello")) &&
-        assertTrue(result.map(_.b) == Right(42)) &&
-        assertTrue(result.map(_.y) == Right(true)) &&
-        assertTrue(result.map(_.e) == Right(0L))
-      },
-      test("build with many shared fields") {
-        // ManySharedSource: shared1, shared2, shared3, removed
-        // ManySharedTarget: shared1, shared2, shared3, added
-        // Only need to handle "removed" and provide "added"
-
-        val migration = MigrationBuilder
-          .newBuilder[ManySharedSource, ManySharedTarget]
-          .dropField(_.removed, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-          .addField(_.added, SchemaExpr.Literal[DynamicValue, Long](0L, Schema.long))
-          .build
-
-        val source = ManySharedSource("a", 1, true, 2.5)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.shared1) == Right("a")) &&
-        assertTrue(result.map(_.shared2) == Right(1)) &&
-        assertTrue(result.map(_.shared3) == Right(true)) &&
-        assertTrue(result.map(_.added) == Right(0L))
-      }
-    ),
-    suite("build compile failures")(
-      test("build fails to compile when drop is missing") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(name: String, extra: Int)
-          case class Tgt(name: String)
-          implicit val srcSchema: Schema[Src] = Schema.derived
-          implicit val tgtSchema: Schema[Tgt] = Schema.derived
-
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Unhandled paths from source")))
-      },
-      test("build fails to compile when add is missing") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(name: String)
-          case class Tgt(name: String, extra: Int)
-          implicit val srcSchema: Schema[Src] = Schema.derived
-          implicit val tgtSchema: Schema[Tgt] = Schema.derived
-
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Unprovided paths for target")))
-      },
-      test("build fails to compile for incomplete complex migration") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(a: String, b: Int, c: Boolean)
-          case class Tgt(x: String, b: Int, y: Boolean)
-          implicit val srcSchema: Schema[Src] = Schema.derived
-          implicit val tgtSchema: Schema[Tgt] = Schema.derived
-
-          // Only handles "a" -> "x", missing c -> y
-          MigrationBuilder.newBuilder[Src, Tgt]
-            .renameField(_.a, _.x)
-            .build
-        """))(Assertion.isLeft)
-      },
-      test("build fails when both handled and provided are missing") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(shared: String, removed: Int)
-          case class Tgt(shared: String, added: Boolean)
-          given Schema[Src] = Schema.derived
-          given Schema[Tgt] = Schema.derived
-
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Unhandled paths from source")))
-      },
-      test("build fails when only some fields are handled") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(keep: String, drop1: Int, drop2: Boolean)
-          case class Tgt(keep: String)
-          given Schema[Src] = Schema.derived
-          given Schema[Tgt] = Schema.derived
-
-          // Empty builder - missing both drop1 and drop2
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Unhandled paths from source")))
-      },
-      test("build fails when only some fields are provided") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(keep: String)
-          case class Tgt(keep: String, add1: Int, add2: Boolean)
-          given Schema[Src] = Schema.derived
-          given Schema[Tgt] = Schema.derived
-
-          // Empty builder - missing both add1 and add2
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Unprovided paths for target")))
-      },
-      test("error message includes field names") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(name: String, fieldToRemove: Int)
-          case class Tgt(name: String)
-          given Schema[Src] = Schema.derived
-          given Schema[Tgt] = Schema.derived
-
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("fieldToRemove")))
-      },
-      test("error message includes type names") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class MySourceType(name: String, extra: Int)
-          case class MyTargetType(name: String)
-          given Schema[MySourceType] = Schema.derived
-          given Schema[MyTargetType] = Schema.derived
-
-          MigrationBuilder.newBuilder[MySourceType, MyTargetType].build
-        """))(Assertion.isLeft(Assertion.containsString("MySourceType")))
-      },
-      test("error message includes hint with example path") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(fieldToRemove: Int)
-          case class Tgt()
-          given Schema[Src] = Schema.derived
-          given Schema[Tgt] = Schema.derived
-
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Hint: Use .dropField(_.fieldToRemove")))
-      },
-      test("error message format is multi-line") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Src(a: Int, b: String)
-          case class Tgt(c: Boolean)
-          given Schema[Src] = Schema.derived
-          given Schema[Tgt] = Schema.derived
-
-          MigrationBuilder.newBuilder[Src, Tgt].build
-        """))(Assertion.isLeft(Assertion.containsString("Unhandled paths from source")))
-      }
-    ),
-    suite("buildPartial always succeeds")(
-      test("buildPartial succeeds even for empty migration on different schemas") {
-        val migration = MigrationBuilder
-          .newBuilder[DropSource, DropTarget]
-          .buildPartial
-
-        // buildPartial doesn't validate, so this should compile
-        assertTrue(migration != null) &&
-        assertTrue(migration.sourceSchema == Schema[DropSource]) &&
-        assertTrue(migration.targetSchema == Schema[DropTarget])
-      },
-      test("buildPartial succeeds for incomplete migration") {
-        val migration = MigrationBuilder
-          .newBuilder[ComplexSource, ComplexTarget]
-          .renameField(_.a, _.x) // Only partial migration
-          .buildPartial
-
-        assertTrue(migration != null)
-      },
-      test("buildPartial works for incomplete drop migration") {
-        case class MultiDropSrc(keep: String, drop1: Int, drop2: Boolean, drop3: Double)
-        case class MultiDropTgt(keep: String)
-
-        given Schema[MultiDropSrc] = Schema.derived
-        given Schema[MultiDropTgt] = Schema.derived
-
-        val partial = MigrationBuilder
-          .newBuilder[MultiDropSrc, MultiDropTgt]
-          .dropField(_.drop1, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-        val migration = partial.buildPartial
-
-        // Only dropped 1 of 3 required fields, but buildPartial doesn't validate
-        assertTrue(migration != null)
-      },
-      test("buildPartial works for incomplete add migration") {
-        case class MultiAddSrc(keep: String)
-        case class MultiAddTgt(keep: String, add1: Int, add2: Boolean, add3: Double)
-
-        given Schema[MultiAddSrc] = Schema.derived
-        given Schema[MultiAddTgt] = Schema.derived
-
-        val partial = MigrationBuilder
-          .newBuilder[MultiAddSrc, MultiAddTgt]
-          .addField(_.add1, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-        val migration = partial.buildPartial
-
-        // Only added 1 of 3 required fields, but buildPartial doesn't validate
-        assertTrue(migration != null)
-      }
-    ),
-    suite("chaining styles")(
-      test("inline chaining works") {
-        val migration = MigrationBuilder
-          .newBuilder[ComplexSource, ComplexTarget]
-          .renameField(_.a, _.x)
-          .renameField(_.c, _.y)
-          .dropField(_.d, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-          .addField(_.e, SchemaExpr.Literal[DynamicValue, Long](0L, Schema.long))
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("multi-line with val assignments works") {
-        val builder1  = MigrationBuilder.newBuilder[ComplexSource, ComplexTarget]
-        val builder2  = builder1.renameField(_.a, _.x)
-        val builder3  = builder2.renameField(_.c, _.y)
-        val builder4  = builder3.dropField(_.d, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-        val builder5  = builder4.addField(_.e, SchemaExpr.Literal[DynamicValue, Long](0L, Schema.long))
-        val migration = builder5.build
-
-        val source = ComplexSource("hello", 42, true, 3.14)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.x) == Right("hello")) &&
-        assertTrue(result.map(_.b) == Right(42)) &&
-        assertTrue(result.map(_.y) == Right(true)) &&
-        assertTrue(result.map(_.e) == Right(0L))
-      },
-      test("mixed inline and multi-line works") {
-        val builder = MigrationBuilder
-          .newBuilder[ComplexSource, ComplexTarget]
-          .renameField(_.a, _.x)
-          .renameField(_.c, _.y)
-
-        val builder2 = builder.dropField(_.d, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-
-        val migration = builder2
-          .addField(_.e, SchemaExpr.Literal[DynamicValue, Long](0L, Schema.long))
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("multiple drops over multiple lines") {
-        case class MultiDropSrc(keep: String, drop1: Int, drop2: Boolean, drop3: Double)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class MultiDropTgt(keep: String)
-
-        given Schema[MultiDropSrc] = Schema.derived
-        given Schema[MultiDropTgt] = Schema.derived
-
-        val b1        = MigrationBuilder.newBuilder[MultiDropSrc, MultiDropTgt]
-        val b2        = b1.dropField(_.drop1, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-        val b3        = b2.dropField(_.drop2, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-        val b4        = b3.dropField(_.drop3, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-        val migration = b4.build
-
-        val source = MultiDropSrc("keep", 1, true, 2.5)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[MultiDropTgt]) &&
-        assertTrue(result.map(_.keep) == Right("keep"))
-      },
-      test("multiple adds over multiple lines") {
-        case class MultiAddSrc(keep: String)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class MultiAddTgt(keep: String, add1: Int, add2: Boolean, add3: Double)
-
-        given Schema[MultiAddSrc] = Schema.derived
-        given Schema[MultiAddTgt] = Schema.derived
-
-        val b1        = MigrationBuilder.newBuilder[MultiAddSrc, MultiAddTgt]
-        val b2        = b1.addField(_.add1, SchemaExpr.Literal[DynamicValue, Int](42, Schema.int))
-        val b3        = b2.addField(_.add2, SchemaExpr.Literal[DynamicValue, Boolean](true, Schema.boolean))
-        val b4        = b3.addField(_.add3, SchemaExpr.Literal[DynamicValue, Double](3.14, Schema.double))
-        val migration = b4.build
-
-        val source = MultiAddSrc("keep")
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[MultiAddTgt]) &&
-        assertTrue(result.map(_.keep) == Right("keep"))
-      },
-      test("multiple drops inline") {
-        case class MultiDropSrc(keep: String, drop1: Int, drop2: Boolean, drop3: Double)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class MultiDropTgt(keep: String)
-
-        given Schema[MultiDropSrc] = Schema.derived
-        given Schema[MultiDropTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[MultiDropSrc, MultiDropTgt]
-          .dropField(_.drop1, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-          .dropField(_.drop2, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          .dropField(_.drop3, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-          .build
-
-        val source = MultiDropSrc("keep", 1, true, 2.5)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.keep) == Right("keep"))
-      },
-      test("multiple adds inline") {
-        case class MultiAddSrc(keep: String)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class MultiAddTgt(keep: String, add1: Int, add2: Boolean, add3: Double)
-
-        given Schema[MultiAddSrc] = Schema.derived
-        given Schema[MultiAddTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[MultiAddSrc, MultiAddTgt]
-          .addField(_.add1, SchemaExpr.Literal[DynamicValue, Int](42, Schema.int))
-          .addField(_.add2, SchemaExpr.Literal[DynamicValue, Boolean](true, Schema.boolean))
-          .addField(_.add3, SchemaExpr.Literal[DynamicValue, Double](3.14, Schema.double))
-          .build
-
-        val source = MultiAddSrc("keep")
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.keep) == Right("keep"))
-      },
-      test("rename handles both source and target tracking") {
-        // Rename should mark oldName as handled and newName as provided
-        val migration = MigrationBuilder
-          .newBuilder[RenameSource, RenameTarget]
-          .renameField(_.oldName, _.newName)
-          .build
-
-        assertTrue(migration != null)
-      }
-    ),
-    suite("edge cases")(
-      test("empty source schema") {
-        case class EmptySrc()
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class NonEmptyTgt(field: String)
-
-        implicit val emptySrcSchema: Schema[EmptySrc]       = Schema.derived
-        implicit val nonEmptyTgtSchema: Schema[NonEmptyTgt] = Schema.derived
-
-        // Only need to provide "field", nothing to handle
-        val migration = MigrationBuilder
-          .newBuilder[EmptySrc, NonEmptyTgt]
-          .addField(_.field, SchemaExpr.Literal[DynamicValue, String]("default", Schema.string))
-          .build
-
-        val result = migration(EmptySrc())
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[NonEmptyTgt])
-      },
-      test("empty target schema") {
-        case class NonEmptySrc(field: String)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class EmptyTgt()
-
-        implicit val nonEmptySrcSchema: Schema[NonEmptySrc] = Schema.derived
-        implicit val emptyTgtSchema: Schema[EmptyTgt]       = Schema.derived
-
-        // Only need to handle "field", nothing to provide
-        val migration = MigrationBuilder
-          .newBuilder[NonEmptySrc, EmptyTgt]
-          .dropField(_.field, SchemaExpr.Literal[DynamicValue, String]("default", Schema.string))
-          .build
-
-        val result = migration(NonEmptySrc("test"))
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[EmptyTgt])
-      },
-      test("empty to empty schema") {
-        case class EmptySource()
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class EmptyTarget()
-
-        given Schema[EmptySource] = Schema.derived
-        given Schema[EmptyTarget] = Schema.derived
-
-        val migration = MigrationBuilder.newBuilder[EmptySource, EmptyTarget].build
-
-        val source = EmptySource()
-        val result = migration(source)
-
-        assertTrue(result.isRight)
-      },
-      test("all fields changed") {
-        case class AllChangedSrc(a: String, b: Int)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class AllChangedTgt(x: Boolean, y: Double)
-
-        implicit val allChangedSrcSchema: Schema[AllChangedSrc] = Schema.derived
-        implicit val allChangedTgtSchema: Schema[AllChangedTgt] = Schema.derived
-
-        // Need to handle a, b and provide x, y
-        val migration = MigrationBuilder
-          .newBuilder[AllChangedSrc, AllChangedTgt]
-          .dropField(_.a, SchemaExpr.Literal[DynamicValue, String]("", Schema.string))
-          .dropField(_.b, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-          .addField(_.x, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          .addField(_.y, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-          .build
-
-        val result = migration(AllChangedSrc("test", 42))
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[AllChangedTgt])
-      },
-      test("single field schemas") {
-        case class SingleA(only: String)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class SingleB(only: String)
-
-        implicit val singleASchema: Schema[SingleA] = Schema.derived
-        implicit val singleBSchema: Schema[SingleB] = Schema.derived
-
-        // Identical schemas - build should compile with no operations
-        val migration = MigrationBuilder.newBuilder[SingleA, SingleB].build
-        val result    = migration(SingleA("test"))
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[SingleB])
-      },
-      test("superset of required handled fields is OK") {
-        case class DropSrc(name: String, age: Int, extra: Boolean)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class DropTgt(name: String, age: Int)
-
-        given Schema[DropSrc] = Schema.derived
-        given Schema[DropTgt] = Schema.derived
-
-        // We can handle more fields than necessary
-        val migration = MigrationBuilder
-          .newBuilder[DropSrc, DropTgt]
-          .dropField(_.extra, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          // Transform a shared field (not required, but allowed)
-          .transformField(_.name, SchemaExpr.Literal[DynamicValue, String]("transformed", Schema.string))
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("superset of required provided fields is OK") {
-        case class AddSrc(name: String, age: Int)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class AddTgt(name: String, age: Int, extra: Boolean)
-
-        given Schema[AddSrc] = Schema.derived
-        given Schema[AddTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[AddSrc, AddTgt]
-          .addField(_.extra, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          // Transform a shared field (not required, but allowed)
-          .transformField(_.name, SchemaExpr.Literal[DynamicValue, String]("transformed", Schema.string))
-          .build
-
-        assertTrue(migration != null)
-      }
-    ),
-    suite("nested path validation")(
-      test("ShapeTree extracts nested structure") {
-        // Verify ShapeTree extracts full nested structure
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class NestedInner(x: Int, y: String)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class NestedOuter(inner: NestedInner, z: Boolean)
-
-        val st = summon[ShapeTree[NestedOuter]]
-        // Verify the tree contains nested RecordNode for "inner"
-        assertTrue(
-          st.tree == ShapeNode.RecordNode(
-            Map(
-              "inner" -> ShapeNode.RecordNode(
-                Map(
-                  "x" -> ShapeNode.PrimitiveNode,
-                  "y" -> ShapeNode.PrimitiveNode
-                )
-              ),
-              "z" -> ShapeNode.PrimitiveNode
-            )
-          )
-        )
-      },
-      test("unchanged nested structure requires no handling") {
-        // When nested structure is identical in both types, no handling needed
-        case class SharedNested(a: Int, b: String)
-        case class TypeWithNestedSrc(shared: SharedNested, remove: Boolean)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class TypeWithNestedTgt(
-          shared: SharedNested,
-          add: Double
-        )
-
-        given Schema[SharedNested]      = Schema.derived
-        given Schema[TypeWithNestedSrc] = Schema.derived
-        given Schema[TypeWithNestedTgt] = Schema.derived
-
-        // The "shared", "shared.a", "shared.b" paths exist in both, so only
-        // "remove" needs to be handled and "add" needs to be provided
-        val migration = MigrationBuilder
-          .newBuilder[TypeWithNestedSrc, TypeWithNestedTgt]
-          .dropField(_.remove, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          .addField(_.add, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-          .build
-
-        val source = TypeWithNestedSrc(SharedNested(42, "hello"), true)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.shared.a) == Right(42)) &&
-        assertTrue(result.map(_.shared.b) == Right("hello")) &&
-        assertTrue(result.map(_.add) == Right(0.0))
-      },
-      test("nested field change requires handling with full path") {
-        // When a nested field changes, the full path must be handled/provided
-        case class AddressV1(street: String, city: String)
-        case class PersonV1(name: String, address: AddressV1)
-
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class AddressV2(street: String, zip: String) // city -> zip
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class PersonV2(name: String, address: AddressV2)
-
-        given Schema[AddressV1] = Schema.derived
-        given Schema[PersonV1]  = Schema.derived
-        given Schema[AddressV2] = Schema.derived
-        given Schema[PersonV2]  = Schema.derived
-
-        // Need to handle "address.city" and provide "address.zip"
-        val migration = MigrationBuilder
-          .newBuilder[PersonV1, PersonV2]
-          .dropField(_.address.city, SchemaExpr.Literal[DynamicValue, String]("", Schema.string))
-          .addField(_.address.zip, SchemaExpr.Literal[DynamicValue, String]("00000", Schema.string))
-          .build
-
-        val source = PersonV1("John", AddressV1("Main St", "NYC"))
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("John")) &&
-        assertTrue(result.map(_.address.street) == Right("Main St")) &&
-        assertTrue(result.map(_.address.zip) == Right("00000"))
-      },
-      test("nested field change fails without full path handling") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class AddressA(street: String, city: String)
-          case class PersonA(name: String, address: AddressA)
-
-          case class AddressB(street: String, zip: String)
-          case class PersonB(name: String, address: AddressB)
-
-          given Schema[AddressA] = Schema.derived
-          given Schema[PersonA]  = Schema.derived
-          given Schema[AddressB] = Schema.derived
-          given Schema[PersonB]  = Schema.derived
-
-          // This should fail - nested change detected but not handled
-          MigrationBuilder.newBuilder[PersonA, PersonB].build
-        """))(Assertion.isLeft(Assertion.containsString("address.city")))
-      },
-      test("nested field rename with full path") {
-        // Rename handles both source and target paths
-        case class AddressV1(street: String, city: String)
-        case class PersonV1(name: String, address: AddressV1)
-
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class AddressV2(street: String, zip: String) // city -> zip
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class PersonV2(name: String, address: AddressV2)
-
-        given Schema[AddressV1] = Schema.derived
-        given Schema[PersonV1]  = Schema.derived
-        given Schema[AddressV2] = Schema.derived
-        given Schema[PersonV2]  = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[PersonV1, PersonV2]
-          .renameField(_.address.city, _.address.zip)
-          .build
-
-        val source = PersonV1("Jane", AddressV1("Oak Ave", "LA"))
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("Jane")) &&
-        assertTrue(result.map(_.address.street) == Right("Oak Ave")) &&
-        assertTrue(result.map(_.address.zip) == Right("LA"))
-      },
-      test("error message shows full nested path for unhandled field") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Inner(old: String)
-          case class Outer(nested: Inner)
-
-          case class InnerNew(newField: String)
-          case class OuterNew(nested: InnerNew)
-
-          given Schema[Inner] = Schema.derived
-          given Schema[Outer] = Schema.derived
-          given Schema[InnerNew] = Schema.derived
-          given Schema[OuterNew] = Schema.derived
-
-          MigrationBuilder.newBuilder[Outer, OuterNew].build
-        """))(Assertion.isLeft(Assertion.containsString("nested.old")))
-      },
-      test("error message shows full nested path for unprovided field") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Inner(x: Int)
-          case class Outer(nested: Inner)
-
-          case class InnerNew(x: Int, y: String)
-          case class OuterNew(nested: InnerNew)
-
-          given Schema[Inner] = Schema.derived
-          given Schema[Outer] = Schema.derived
-          given Schema[InnerNew] = Schema.derived
-          given Schema[OuterNew] = Schema.derived
-
-          MigrationBuilder.newBuilder[Outer, OuterNew].build
-        """))(Assertion.isLeft(Assertion.containsString("nested.y")))
-      },
-      test("deeply nested path tracking (3 levels)") {
-        case class Level3(value: String)
-        case class Level2(l3: Level3)
-        case class Level1Src(l2: Level2, extra: Int)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class Level1Tgt(l2: Level2)
-
-        given Schema[Level3]    = Schema.derived
-        given Schema[Level2]    = Schema.derived
-        given Schema[Level1Src] = Schema.derived
-        given Schema[Level1Tgt] = Schema.derived
-
-        // Nested l2, l2.l3, l2.l3.value are shared, only "extra" needs handling
-        val migration = MigrationBuilder
-          .newBuilder[Level1Src, Level1Tgt]
-          .dropField(_.extra, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-          .build
-
-        val source = Level1Src(Level2(Level3("deep")), 42)
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.l2.l3.value) == Right("deep"))
-      },
-      test("deeply nested change requires full path") {
-        case class DeepInnerV1(value: Int)
-        case class DeepMiddleV1(inner: DeepInnerV1)
-        case class DeepOuterV1(middle: DeepMiddleV1)
-
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class DeepInnerV2(value: Long) // Int -> Long
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class DeepMiddleV2(inner: DeepInnerV2)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class DeepOuterV2(middle: DeepMiddleV2)
-
-        given Schema[DeepInnerV1]  = Schema.derived
-        given Schema[DeepMiddleV1] = Schema.derived
-        given Schema[DeepOuterV1]  = Schema.derived
-        given Schema[DeepInnerV2]  = Schema.derived
-        given Schema[DeepMiddleV2] = Schema.derived
-        given Schema[DeepOuterV2]  = Schema.derived
-
-        // The change at middle.inner.value needs to be handled
-        val migration = MigrationBuilder
-          .newBuilder[DeepOuterV1, DeepOuterV2]
-          .changeFieldType(_.middle.inner.value, PrimitiveConverter.IntToLong)
-          .build
-
-        val source = DeepOuterV1(DeepMiddleV1(DeepInnerV1(42)))
-        val result = migration(source)
-
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.middle.inner.value) == Right(42L))
-      }
-    ),
-    suite("sealed trait case validation")(
-      test("ShapeTree extracts cases for sealed traits") {
-        // Verify ShapeTree extracts case names for sealed traits as SealedNode
-        sealed trait SimpleResult
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class Ok(value: Int) extends SimpleResult
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class Err(msg: String) extends SimpleResult
-
-        val st = summon[ShapeTree[SimpleResult]]
-        // The tree should be a SealedNode with "Err" and "Ok" cases
-        assertTrue(
-          st.tree == ShapeNode.SealedNode(
-            Map(
-              "Err" -> ShapeNode.RecordNode(Map("msg" -> ShapeNode.PrimitiveNode)),
-              "Ok"  -> ShapeNode.RecordNode(Map("value" -> ShapeNode.PrimitiveNode))
-            )
-          )
-        )
-      },
-      test("ShapeTree returns RecordNode for non-sealed types") {
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class RegularClass(x: Int)
-
-        val st = summon[ShapeTree[RegularClass]]
-        assertTrue(st.tree == ShapeNode.RecordNode(Map("x" -> ShapeNode.PrimitiveNode)))
-      },
-      test("different case names require handling") {
-        // When sealed traits have different case names, handling is required
-        // Verify that build fails without handling case renames
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          sealed trait StatusV1
-          case class Active(since: String) extends StatusV1
-          case class Inactive() extends StatusV1
-
-          sealed trait StatusV2
-          case class Active2(since: String) extends StatusV2
-          case class Inactive2() extends StatusV2
-
-          given Schema[StatusV1] = Schema.derived
-          given Schema[StatusV2] = Schema.derived
-
-          // This should fail - case names changed but not handled
-          MigrationBuilder.newBuilder[StatusV1, StatusV2].build
-        """))(Assertion.isLeft)
-      },
-      test("case rename requires handling") {
-        // When a case is renamed, it must be explicitly handled
-        // Verify that build fails without handling case renames
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          sealed trait PaymentV1
-          case class CardPayment(number: String) extends PaymentV1
-          case class CashPayment(amount: Int) extends PaymentV1
-
-          sealed trait PaymentV2
-          case class CreditCard(number: String) extends PaymentV2
-          case class CashPayment2(amount: Int) extends PaymentV2
-
-          given Schema[PaymentV1] = Schema.derived
-          given Schema[PaymentV2] = Schema.derived
-
-          // This should fail - case renames not handled
-          MigrationBuilder.newBuilder[PaymentV1, PaymentV2].build
-        """))(Assertion.isLeft)
-      },
-      test("renameCase tracks case in Handled and Provided") {
-        sealed trait ResultV1
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class SuccessV1(value: Int) extends ResultV1
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class FailureV1(err: String) extends ResultV1
-        object ResultV1                   extends CompanionOptics[ResultV1]
-
-        sealed trait ResultV2
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class OkResult(value: Int) extends ResultV2 // Renamed from SuccessV1
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class ErrResult(err: String) extends ResultV2 // Renamed from FailureV1
-
-        given Schema[ResultV1]  = Schema.derived
-        given Schema[ResultV2]  = Schema.derived
-        given Schema[SuccessV1] = Schema.derived
-        given Schema[FailureV1] = Schema.derived
-        given Schema[OkResult]  = Schema.derived
-        given Schema[ErrResult] = Schema.derived
-
-        import ResultV1.when
-
-        val migration = MigrationBuilder
-          .newBuilder[ResultV1, ResultV2]
-          .renameCase((r: ResultV1) => r.when[SuccessV1], "OkResult")
-          .renameCase((r: ResultV1) => r.when[FailureV1], "ErrResult")
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("missing case handling fails to compile") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          sealed trait OldTrait
-          case class CaseA(x: Int) extends OldTrait
-          case class CaseB(y: String) extends OldTrait
-
-          sealed trait NewTrait
-          case class CaseX(x: Int) extends NewTrait    // renamed from CaseA
-          case class CaseB2(y: String) extends NewTrait // renamed from CaseB
-
-          given Schema[OldTrait] = Schema.derived
-          given Schema[NewTrait] = Schema.derived
-          given Schema[CaseA]    = Schema.derived
-          given Schema[CaseB]    = Schema.derived
-          given Schema[CaseX]    = Schema.derived
-          given Schema[CaseB2]   = Schema.derived
-
-          // This should fail - case names changed but not handled
-          MigrationBuilder.newBuilder[OldTrait, NewTrait].build
-        """))(Assertion.isLeft(Assertion.containsString("Unhandled cases from source")))
-      },
-      test("error message shows unhandled case names") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          sealed trait PaymentV1
-          case class CreditCard(number: String) extends PaymentV1
-          case class Cash(amount: Int) extends PaymentV1
-
-          sealed trait PaymentV2
-          case class Card(number: String) extends PaymentV2
-          case class Money(amount: Int) extends PaymentV2
-
-          given Schema[PaymentV1] = Schema.derived
-          given Schema[PaymentV2] = Schema.derived
-          given Schema[CreditCard] = Schema.derived
-          given Schema[Cash] = Schema.derived
-          given Schema[Card] = Schema.derived
-          given Schema[Money] = Schema.derived
-
-          MigrationBuilder.newBuilder[PaymentV1, PaymentV2].build
-        """))(Assertion.isLeft(Assertion.containsString("CreditCard") && Assertion.containsString("Cash")))
-      },
-      test("error message shows unprovided case names") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          sealed trait StatusV1
-          case class Active(since: String) extends StatusV1
-
-          sealed trait StatusV2
-          case class Active2(since: String) extends StatusV2
-          case class Disabled(reason: String) extends StatusV2
-
-          given Schema[StatusV1] = Schema.derived
-          given Schema[StatusV2] = Schema.derived
-          given Schema[Active] = Schema.derived
-          given Schema[Active2] = Schema.derived
-          given Schema[Disabled] = Schema.derived
-
-          MigrationBuilder.newBuilder[StatusV1, StatusV2].build
-        """))(Assertion.isLeft(Assertion.containsString("Unprovided cases for target")))
-      },
-      test("identical sealed traits require no case handling") {
-        sealed trait StatusSame
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class ActiveSame(since: String) extends StatusSame
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class InactiveSame() extends StatusSame
-
-        given Schema[StatusSame] = Schema.derived
-
-        val migration = MigrationBuilder.newBuilder[StatusSame, StatusSame].build
-
-        assertTrue(migration != null)
-      },
-      test("sealed trait with case objects") {
-        sealed trait WithCaseObject
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class SomeValue(x: Int) extends WithCaseObject
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object NoneValue extends WithCaseObject
-
-        given Schema[WithCaseObject] = Schema.derived
-
-        // ShapeTree should include both case classes and case objects as SealedNode
-        val st = summon[ShapeTree[WithCaseObject]]
-        // Should have "NoneValue" and "SomeValue" cases
-        assertTrue(
-          st.tree == ShapeNode.SealedNode(
-            Map(
-              "SomeValue" -> ShapeNode.RecordNode(Map("x" -> ShapeNode.PrimitiveNode)),
-              "NoneValue" -> ShapeNode.RecordNode(Map.empty)
-            )
-          )
-        )
-      },
-      test("enum-like sealed trait with only case objects") {
-        sealed trait Color
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object Red extends Color
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object Green extends Color
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object Blue extends Color
-
-        given Schema[Color] = Schema.derived
-
-        // All case objects should be extracted
-        val st = summon[ShapeTree[Color]]
-        // Should have "Blue", "Green", "Red" cases
-        assertTrue(
-          st.tree == ShapeNode.SealedNode(
-            Map(
-              "Red"   -> ShapeNode.RecordNode(Map.empty),
-              "Green" -> ShapeNode.RecordNode(Map.empty),
-              "Blue"  -> ShapeNode.RecordNode(Map.empty)
-            )
-          )
-        )
-      },
-      test("identical sealed traits with case objects require no handling") {
-        sealed trait WithCaseObject
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class SomeValue(x: Int) extends WithCaseObject
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object NoneValue extends WithCaseObject
-
-        given Schema[WithCaseObject] = Schema.derived
-
-        // Same case object structure should compile without handling
-        val migration = MigrationBuilder.newBuilder[WithCaseObject, WithCaseObject].build
-        assertTrue(migration != null)
-      },
-      test("identical enum-like sealed traits require no handling") {
-        sealed trait Color
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object Red extends Color
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object Green extends Color
-        @scala.annotation.nowarn("msg=unused local definition")
-        case object Blue extends Color
-
-        given Schema[Color] = Schema.derived
-
-        // Same enum structure should compile without handling
-        val migration = MigrationBuilder.newBuilder[Color, Color].build
-        assertTrue(migration != null)
-      },
-      test("partial case handling fails to compile") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          sealed trait SourceTrait
-          case class Alpha(x: Int) extends SourceTrait
-          case class Beta(y: String) extends SourceTrait
-
-          sealed trait TargetTrait
-          case class AlphaNew(x: Int) extends TargetTrait
-          case class BetaNew(y: String) extends TargetTrait
-
-          given Schema[SourceTrait] = Schema.derived
-          given Schema[TargetTrait] = Schema.derived
-          given Schema[Alpha]       = Schema.derived
-          given Schema[Beta]        = Schema.derived
-          given Schema[AlphaNew]    = Schema.derived
-          given Schema[BetaNew]     = Schema.derived
-
-          // Only handle one case rename - should fail
-          MigrationBuilder.newBuilder[SourceTrait, TargetTrait]
-            .renameCase(_.when[Alpha], "AlphaNew")
-            .build
-        """))(Assertion.isLeft)
-      },
+/**
+ * Scala 3-specific CompileTimeValidationSpec tests.
+ *
+ * Tests Scala 3 `enum` support and `given` syntax which are not available in
+ * Scala 2.
+ *
+ * All shared tests (direct method chaining, typeCheck, ShapeTree, etc.) are in
+ * `src/test/scala/zio/blocks/schema/migration/CompileTimeValidationSpec.scala`.
+ */
+object CompileTimeValidationSpecScala3 extends ZIOSpecDefault {
+
+  def spec = suite("CompileTimeValidationSpec - Scala 3 specific")(
+    suite("enum support")(
       test("enum case rename works") {
         enum ColorV1 {
           case Red, Green, Blue
@@ -1125,487 +38,30 @@ object CompileTimeValidationSpec extends ZIOSpecDefault {
           .renameCase((c: ColorV1) => c.when[ColorV1.Red.type], "Crimson")
           .build
 
-        assertTrue(migration != null)
-      },
-      test("mixed field and case changes") {
-        // When both fields and cases change, both must be handled
-        // Verify that build fails without handling
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
+        // Verify build succeeds and the rename works at DynamicValue level
+        val sourceSchema = summon[Schema[ColorV1]]
+        val dynamicRed   = sourceSchema.toDynamicValue(ColorV1.Red)
+        val result       = migration.dynamicMigration.apply(dynamicRed)
 
-          sealed trait ContainerV1
-          case class BoxV1(value: Int, label: String) extends ContainerV1
-
-          sealed trait ContainerV2
-          case class BoxV2(value: Int) extends ContainerV2 // Renamed case, removed field
-
-          given Schema[ContainerV1] = Schema.derived
-          given Schema[ContainerV2] = Schema.derived
-
-          // This should fail - case rename and field changes not handled
-          MigrationBuilder.newBuilder[ContainerV1, ContainerV2].build
-        """))(Assertion.isLeft)
-      }
-    ),
-    suite("cross-path compile-time validation")(
-      test("joinFields should fail at compile time when source paths have different parents") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema._
-          import zio.blocks.schema.migration._
-
-          case class Source(meta: Meta, data: Data)
-          case class Meta(id: String)
-          case class Data(value: Int)
-          case class Target(result: String)
-
-          given Schema[Meta] = Schema.derived
-          given Schema[Data] = Schema.derived
-          given Schema[Source] = Schema.derived
-          given Schema[Target] = Schema.derived
-
-          MigrationBuilder.newBuilder[Source, Target]
-            .joinFields(
-              (t: Target) => t.result,
-              Seq((s: Source) => s.meta.id, (s: Source) => s.data.value),
-              SchemaExpr.Literal[DynamicValue, String]("combined", Schema.string)
-            )
-        """))(Assertion.isLeft(Assertion.containsString("joinFields source fields must share common parent")))
-      },
-      test("splitField should fail at compile time when target paths have different parents") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema._
-          import zio.blocks.schema.migration._
-
-          case class Source(data: String)
-          case class Target(meta: Meta, info: Info)
-          case class Meta(part1: String)
-          case class Info(part2: String)
-
-          given Schema[Meta] = Schema.derived
-          given Schema[Info] = Schema.derived
-          given Schema[Source] = Schema.derived
-          given Schema[Target] = Schema.derived
-
-          MigrationBuilder.newBuilder[Source, Target]
-            .splitField(
-              (s: Source) => s.data,
-              Seq((t: Target) => t.meta.part1, (t: Target) => t.info.part2),
-              SchemaExpr.Literal[DynamicValue, String]("split", Schema.string)
-            )
-        """))(Assertion.isLeft(Assertion.containsString("splitField target fields must share common parent")))
-      },
-      test("joinFields should compile when source paths share common parent") {
-        case class NameSrc(first: String, last: String)
-        case class NameTgt(full: String)
-        case class PersonSrc(name: NameSrc)
-        case class PersonTgt(name: NameTgt)
-
-        given Schema[NameSrc]   = Schema.derived
-        given Schema[NameTgt]   = Schema.derived
-        given Schema[PersonSrc] = Schema.derived
-        given Schema[PersonTgt] = Schema.derived
-
-        // This should compile since both paths share parent "name"
-        // name.first and name.last are joined into name.full
-        val migration = MigrationBuilder
-          .newBuilder[PersonSrc, PersonTgt]
-          .joinFields(
-            (t: PersonTgt) => t.name.full,
-            Seq((s: PersonSrc) => s.name.first, (s: PersonSrc) => s.name.last),
-            SchemaExpr.Literal[DynamicValue, String]("combined", Schema.string)
-          )
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("splitField should compile when target paths share common parent") {
-        case class NameSrc(full: String)
-        case class NameTgt(first: String, last: String)
-        case class PersonSrc(name: NameSrc)
-        case class PersonTgt(name: NameTgt)
-
-        given Schema[NameSrc]   = Schema.derived
-        given Schema[NameTgt]   = Schema.derived
-        given Schema[PersonSrc] = Schema.derived
-        given Schema[PersonTgt] = Schema.derived
-
-        // This should compile since both paths share parent "name"
-        // name.full is split into name.first and name.last
-        val migration = MigrationBuilder
-          .newBuilder[PersonSrc, PersonTgt]
-          .splitField(
-            (s: PersonSrc) => s.name.full,
-            Seq((t: PersonTgt) => t.name.first, (t: PersonTgt) => t.name.last),
-            SchemaExpr.Literal[DynamicValue, String]("split", Schema.string)
-          )
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("joinFields with single source path should compile") {
-        // Uses existing FullNameSource/FullNameTarget, drops lastName to complete migration
-        val migration = MigrationBuilder
-          .newBuilder[FullNameSource, FullNameTarget]
-          .joinFields(
-            (t: FullNameTarget) => t.fullName,
-            Seq((s: FullNameSource) => s.firstName),
-            SchemaExpr.Literal[DynamicValue, String]("value", Schema.string)
-          )
-          .dropField(
-            (s: FullNameSource) => s.lastName,
-            SchemaExpr.Literal[DynamicValue, String]("", Schema.string)
-          )
-          .build
-
-        assertTrue(migration != null)
-      },
-      test("splitField with single target path should compile") {
-        // Uses existing FullNameTarget/FullNameSource, adds lastName to complete migration
-        val migration = MigrationBuilder
-          .newBuilder[FullNameTarget, FullNameSource]
-          .splitField(
-            (s: FullNameTarget) => s.fullName,
-            Seq((t: FullNameSource) => t.firstName),
-            SchemaExpr.Literal[DynamicValue, String]("value", Schema.string)
-          )
-          .addField(
-            (t: FullNameSource) => t.lastName,
-            SchemaExpr.Literal[DynamicValue, String]("", Schema.string)
-          )
-          .build
-
-        assertTrue(migration != null)
-      }
-    ),
-    suite("joinFields and splitField tracking")(
-      test("joinFields tracks all source fields in Handled") {
-        // joinFields should handle firstName and lastName, and provide fullName
-        val builder   = MigrationBuilder.newBuilder[FullNameSource, FullNameTarget]
-        val afterJoin = builder.joinFields(
-          (t: FullNameTarget) => t.fullName,
-          Seq((s: FullNameSource) => s.firstName, (s: FullNameSource) => s.lastName),
-          SchemaExpr.Literal[DynamicValue, String]("combined", Schema.string)
-        )
-        // If joinFields didn't track firstName and lastName as Handled, .build wouldn't compile
-        val migration = afterJoin.build
-        assertTrue(migration != null)
-      },
-      test("splitField tracks all target fields in Provided") {
-        // splitField should handle fullName and provide firstName and lastName
-        // Using top-level case classes FullNameTarget (as source) and FullNameSource (as target)
-        val builder    = MigrationBuilder.newBuilder[FullNameTarget, FullNameSource]
-        val afterSplit = builder.splitField(
-          (s: FullNameTarget) => s.fullName,
-          Seq((t: FullNameSource) => t.firstName, (t: FullNameSource) => t.lastName),
-          SchemaExpr.Literal[DynamicValue, String]("split", Schema.string)
-        )
-        // If splitField didn't track firstName and lastName as Provided, .build wouldn't compile
-        val migration = afterSplit.build
-        assertTrue(migration != null)
-      },
-      test("joinFields without tracking all sources fails to compile") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Source(a: String, b: String, shared: Int)
-          case class Target(combined: String, shared: Int)
-
-          given Schema[Source] = Schema.derived
-          given Schema[Target] = Schema.derived
-
-          // Without joinFields handling a and b, this should fail
-          MigrationBuilder.newBuilder[Source, Target]
-            .addField(_.combined, SchemaExpr.Literal[DynamicValue, String]("default", Schema.string))
-            .build
-        """))(Assertion.isLeft)
-      },
-      test("splitField without tracking all targets fails to compile") {
-        assertZIO(typeCheck("""
-          import zio.blocks.schema.migration._
-          import zio.blocks.schema._
-
-          case class Source(combined: String, shared: Int)
-          case class Target(a: String, b: String, shared: Int)
-
-          given Schema[Source] = Schema.derived
-          given Schema[Target] = Schema.derived
-
-          // Without splitField providing a and b, this should fail
-          MigrationBuilder.newBuilder[Source, Target]
-            .dropField(_.combined, SchemaExpr.Literal[DynamicValue, String]("default", Schema.string))
-            .build
-        """))(Assertion.isLeft)
-      }
-    ),
-    suite("edge cases with container types")(
-      test("Option field migration works") {
-        case class WithOptionSrc(name: String, data: Option[String])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class WithOptionTgt(name: String, data: Option[String], extra: Int)
-
-        given Schema[WithOptionSrc] = Schema.derived
-        given Schema[WithOptionTgt] = Schema.derived
-
-        // Only "extra" needs to be provided
-        val migration = MigrationBuilder
-          .newBuilder[WithOptionSrc, WithOptionTgt]
-          .addField(_.extra, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-          .build
-
-        val result = migration(WithOptionSrc("test", Some("value")))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("test")) &&
-        assertTrue(result.map(_.data) == Right(Some("value"))) &&
-        assertTrue(result.map(_.extra) == Right(0))
-      },
-      test("List field migration works") {
-        case class WithListSrc(name: String, items: List[String])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class WithListTgt(name: String, items: List[String], count: Int)
-
-        given Schema[WithListSrc] = Schema.derived
-        given Schema[WithListTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[WithListSrc, WithListTgt]
-          .addField(_.count, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-          .build
-
-        val result = migration(WithListSrc("test", List("a", "b")))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("test")) &&
-        assertTrue(result.map(_.items) == Right(List("a", "b"))) &&
-        assertTrue(result.map(_.count) == Right(0))
-      },
-      test("Set field migration works") {
-        case class WithSetSrc(name: String, tags: Set[String])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class WithSetTgt(name: String, tags: Set[String], active: Boolean)
-
-        given Schema[WithSetSrc] = Schema.derived
-        given Schema[WithSetTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[WithSetSrc, WithSetTgt]
-          .addField(_.active, SchemaExpr.Literal[DynamicValue, Boolean](true, Schema.boolean))
-          .build
-
-        val result = migration(WithSetSrc("test", Set("tag1", "tag2")))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("test")) &&
-        assertTrue(result.map(_.tags) == Right(Set("tag1", "tag2"))) &&
-        assertTrue(result.map(_.active) == Right(true))
-      },
-      test("Map field migration works") {
-        case class WithMapSrc(name: String, data: Map[String, Int])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class WithMapTgt(name: String, data: Map[String, Int], version: Long)
-
-        given Schema[WithMapSrc] = Schema.derived
-        given Schema[WithMapTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[WithMapSrc, WithMapTgt]
-          .addField(_.version, SchemaExpr.Literal[DynamicValue, Long](1L, Schema.long))
-          .build
-
-        val result = migration(WithMapSrc("test", Map("a" -> 1)))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("test")) &&
-        assertTrue(result.map(_.data) == Right(Map("a" -> 1))) &&
-        assertTrue(result.map(_.version) == Right(1L))
-      },
-      test("Vector field migration works") {
-        case class WithVectorSrc(name: String, values: Vector[Double])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class WithVectorTgt(name: String, values: Vector[Double], sum: Double)
-
-        given Schema[WithVectorSrc] = Schema.derived
-        given Schema[WithVectorTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[WithVectorSrc, WithVectorTgt]
-          .addField(_.sum, SchemaExpr.Literal[DynamicValue, Double](0.0, Schema.double))
-          .build
-
-        val result = migration(WithVectorSrc("test", Vector(1.0, 2.0)))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("test")) &&
-        assertTrue(result.map(_.values) == Right(Vector(1.0, 2.0))) &&
-        assertTrue(result.map(_.sum) == Right(0.0))
-      },
-      test("nested case class in Option migration works") {
-        case class Inner(x: Int, y: String)
-        case class OuterSrc(name: String, inner: Option[Inner])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class OuterTgt(name: String, inner: Option[Inner], extra: Boolean)
-
-        given Schema[Inner]    = Schema.derived
-        given Schema[OuterSrc] = Schema.derived
-        given Schema[OuterTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[OuterSrc, OuterTgt]
-          .addField(_.extra, SchemaExpr.Literal[DynamicValue, Boolean](false, Schema.boolean))
-          .build
-
-        val result = migration(OuterSrc("test", Some(Inner(1, "hello"))))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.name) == Right("test")) &&
-        assertTrue(result.map(_.inner.map(_.x)) == Right(Some(1))) &&
-        assertTrue(result.map(_.inner.map(_.y)) == Right(Some("hello"))) &&
-        assertTrue(result.map(_.extra) == Right(false))
-      },
-      test("nested case class in List migration works") {
-        case class Item(id: Int, name: String)
-        case class ContainerSrc(items: List[Item])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class ContainerTgt(items: List[Item], count: Int)
-
-        given Schema[Item]         = Schema.derived
-        given Schema[ContainerSrc] = Schema.derived
-        given Schema[ContainerTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[ContainerSrc, ContainerTgt]
-          .addField(_.count, SchemaExpr.Literal[DynamicValue, Int](0, Schema.int))
-          .build
-
-        val result = migration(ContainerSrc(List(Item(1, "a"), Item(2, "b"))))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.items.map(_.id)) == Right(List(1, 2))) &&
-        assertTrue(result.map(_.items.map(_.name)) == Right(List("a", "b"))) &&
-        assertTrue(result.map(_.count) == Right(0))
-      },
-      test("empty case class ShapeTree is empty RecordNode") {
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class EmptyClass()
-
-        val st = summon[ShapeTree[EmptyClass]]
-        assertTrue(st.tree == ShapeNode.RecordNode(Map.empty))
-      },
-      test("single field case class ShapeTree has one field") {
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class SingleField(value: String)
-
-        val st = summon[ShapeTree[SingleField]]
-        assertTrue(st.tree == ShapeNode.RecordNode(Map("value" -> ShapeNode.PrimitiveNode)))
-      },
-      test("deeply nested containers migration works") {
-        // Nested containers: Option[List[Map[String, Int]]]
-        case class DeepContainerSrc(data: Option[List[Map[String, Int]]])
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class DeepContainerTgt(data: Option[List[Map[String, Int]]], meta: String)
-
-        given Schema[DeepContainerSrc] = Schema.derived
-        given Schema[DeepContainerTgt] = Schema.derived
-
-        val migration = MigrationBuilder
-          .newBuilder[DeepContainerSrc, DeepContainerTgt]
-          .addField(_.meta, SchemaExpr.Literal[DynamicValue, String]("", Schema.string))
-          .build
-
-        val result = migration(DeepContainerSrc(Some(List(Map("a" -> 1)))))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.data) == Right(Some(List(Map("a" -> 1))))) &&
-        assertTrue(result.map(_.meta) == Right(""))
-      },
-      test("container types extract with ShapeTree") {
-        case class AddressV1(street: String, city: String)
-        case class WithContainers(
-          opt: Option[AddressV1],
-          list: List[String],
-          set: Set[Int],
-          map: Map[String, AddressV1]
-        )
-
-        given Schema[AddressV1]      = Schema.derived
-        given Schema[WithContainers] = Schema.derived
-
-        // Container fields should have proper ShapeNode representation
-        val st = summon[ShapeTree[WithContainers]]
-        // ShapeTree should show SeqNode, OptionNode, MapNode for container fields
         assertTrue(
-          st.tree == ShapeNode.RecordNode(
-            Map(
-              "opt" -> ShapeNode.OptionNode(
-                ShapeNode.RecordNode(Map("street" -> ShapeNode.PrimitiveNode, "city" -> ShapeNode.PrimitiveNode))
-              ),
-              "list" -> ShapeNode.SeqNode(ShapeNode.PrimitiveNode),
-              "set"  -> ShapeNode.SeqNode(ShapeNode.PrimitiveNode),
-              "map"  -> ShapeNode.MapNode(
-                ShapeNode.PrimitiveNode,
-                ShapeNode.RecordNode(Map("street" -> ShapeNode.PrimitiveNode, "city" -> ShapeNode.PrimitiveNode))
-              )
-            )
-          )
+          result.isRight,
+          result.exists(_.caseName == Some("Crimson"))
         )
-      },
-      test("deeply nested containers extract correctly") {
-        // List[Option[Map[String, Vector[Int]]]] should extract as nested ShapeNodes
-        case class DeeplyNestedContainers(items: List[Option[Map[String, Vector[Int]]]])
+      }
+    ),
+    suite("given syntax in typeCheck")(
+      test("build fails with given syntax when drop is missing") {
+        assertZIO(typeCheck("""
+          import zio.blocks.schema.migration._
+          import zio.blocks.schema._
 
-        given Schema[DeeplyNestedContainers] = Schema.derived
+          case class Src(shared: String, removed: Int)
+          case class Tgt(shared: String, added: Boolean)
+          given Schema[Src] = Schema.derived
+          given Schema[Tgt] = Schema.derived
 
-        val st = summon[ShapeTree[DeeplyNestedContainers]]
-        assertTrue(
-          st.tree == ShapeNode.RecordNode(
-            Map(
-              "items" -> ShapeNode.SeqNode(
-                ShapeNode.OptionNode(
-                  ShapeNode.MapNode(
-                    ShapeNode.PrimitiveNode,
-                    ShapeNode.SeqNode(ShapeNode.PrimitiveNode)
-                  )
-                )
-              )
-            )
-          )
-        )
-      },
-      test("migration with container fields works") {
-        case class AddressV1(street: String, city: String)
-        case class WithContainers(
-          opt: Option[AddressV1],
-          list: List[String],
-          set: Set[Int],
-          map: Map[String, AddressV1]
-        )
-
-        given Schema[AddressV1]      = Schema.derived
-        given Schema[WithContainers] = Schema.derived
-
-        // Containers with same element types should still migrate if container field names match
-        val migration = MigrationBuilder.newBuilder[WithContainers, WithContainers].build
-        assertTrue(migration != null)
-      },
-      test("empty case class migration works") {
-        case class EmptySource()
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class EmptyTarget()
-
-        given Schema[EmptySource] = Schema.derived
-        given Schema[EmptyTarget] = Schema.derived
-
-        val migration = MigrationBuilder.newBuilder[EmptySource, EmptyTarget].build
-        val result    = migration(EmptySource())
-        assertTrue(result.isRight) &&
-        assertTrue(result.toOption.get.isInstanceOf[EmptyTarget])
-      },
-      test("single field case class migration works") {
-        case class SingleA(only: String)
-        @scala.annotation.nowarn("msg=unused local definition")
-        case class SingleB(only: String)
-
-        given Schema[SingleA] = Schema.derived
-        given Schema[SingleB] = Schema.derived
-
-        val migration = MigrationBuilder.newBuilder[SingleA, SingleB].build
-        val result    = migration(SingleA("test"))
-        assertTrue(result.isRight) &&
-        assertTrue(result.map(_.only) == Right("test"))
+          MigrationBuilder.newBuilder[Src, Tgt].build
+        """))(Assertion.isLeft(Assertion.containsString("Unhandled paths from source")))
       }
     )
   )
